fix(scalars): ListScalar carries its own dtype (V0)
`ListScalar` backs `list`, `large_list`, `map` and `fixed_size_list`, and
rebuilt its type as `list_(child.dtype())`. That reconstruction reports the
*shape*, not the type: a `MapArray` element answered
`list<struct<key, value>>`, a `LargeListArray` element answered `list<...>`,
and a `FixedSizeListArray` element lost its size.

Carry a `_dtype: DynType` instead, set from the array's own dtype by
`ListLikeArray.__getitem__` and `FixedSizeListArray.__getitem__`. One field
fixes all three cases. The cheaper 1-byte kind tag was rejected: rebuilding
`list_`/`large_list_`/`map_` from the child cannot carry `keysSorted` or the
entries field name.

Three of the four new cases fail before the change.

Costs +2,820 bytes of `__text` (+0.136%) on `query_streaming_agg_fused`,
which is why it lands with the size-gate re-baseline.

+++ marrow/arrays.mojo
@@ -1242,7 +1242,9 @@ struct ListLikeArray[T: ListLikeType](Array):
         if index < 0 or index >= self.length:
             raise Error(t"index {index} out of bounds for length {self.length}")
         return ListScalar(
-            value=self.unsafe_get(index), is_valid=self.is_valid(index)
+            dtype=self.dtype,
+            value=self.unsafe_get(index),
+            is_valid=self.is_valid(index),
         )
 
     def slice(self, offset: Int = 0, length: Int = -1) -> Self:
@@ -1537,7 +1539,9 @@ struct FixedSizeListArray(Array):
         if index < 0 or index >= self.length:
             raise Error(t"index {index} out of bounds for length {self.length}")
         return ListScalar(
-            value=self.unsafe_get(index), is_valid=self.is_valid(index)
+            dtype=self.dtype,
+            value=self.unsafe_get(index),
+            is_valid=self.is_valid(index),
         )
 
     def slice(self, offset: Int = 0, length: Int = -1) -> Self:
